// src/lib/api/models/ApiFavoritesGet200ResponseInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kinnema
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiFavoritesGet200ResponseInner
 */
export interface ApiFavoritesGet200ResponseInner {
    /**
     * 
     * @type {string}
     * @memberof ApiFavoritesGet200ResponseInner
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ApiFavoritesGet200ResponseInner
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ApiFavoritesGet200ResponseInner
     */
    posterPath: string;
    /**
     * 
     * @type {number}
     * @memberof ApiFavoritesGet200ResponseInner
     */
    tmdbId: number;
    /**
     * 
     * @type {string}
     * @memberof ApiFavoritesGet200ResponseInner
     */
    userId: string;
    /**
     * 
     * @type {Date}
     * @memberof ApiFavoritesGet200ResponseInner
     */
    createdAt: Date;
}

/**
 * Check if a given object implements the ApiFavoritesGet200ResponseInner interface.
 */
export function instanceOfApiFavoritesGet200ResponseInner(value: object): value is ApiFavoritesGet200ResponseInner {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('posterPath' in value) || value['posterPath'] === undefined) return false;
    if (!('tmdbId' in value) || value['tmdbId'] === undefined) return false;
    if (!('userId' in value) || value['userId'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function ApiFavoritesGet200ResponseInnerFromJSON(json: any): ApiFavoritesGet200ResponseInner {
    return ApiFavoritesGet200ResponseInnerFromJSONTyped(json, false);
}

export function ApiFavoritesGet200ResponseInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiFavoritesGet200ResponseInner {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'posterPath': json['poster_path'],
        'tmdbId': json['tmdbId'],
        'userId': json['userId'],
        'createdAt': (new Date(json['createdAt'])),
    };
}

export function ApiFavoritesGet200ResponseInnerToJSON(json: any): ApiFavoritesGet200ResponseInner {
    return ApiFavoritesGet200ResponseInnerToJSONTyped(json, false);
}

export function ApiFavoritesGet200ResponseInnerToJSONTyped(value?: ApiFavoritesGet200ResponseInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'poster_path': value['posterPath'],
        'tmdbId': value['tmdbId'],
        'userId': value['userId'],
        'createdAt': ((value['createdAt']).toISOString()),
    };
}