// src/lib/api/models/ApiFavoritesPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kinnema
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiFavoritesPostRequest
 */
export interface ApiFavoritesPostRequest {
    /**
     * 
     * @type {string}
     * @memberof ApiFavoritesPostRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ApiFavoritesPostRequest
     */
    posterPath: string;
    /**
     * 
     * @type {number}
     * @memberof ApiFavoritesPostRequest
     */
    tmdbId: number;
}

/**
 * Check if a given object implements the ApiFavoritesPostRequest interface.
 */
export function instanceOfApiFavoritesPostRequest(value: object): value is ApiFavoritesPostRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('posterPath' in value) || value['posterPath'] === undefined) return false;
    if (!('tmdbId' in value) || value['tmdbId'] === undefined) return false;
    return true;
}

export function ApiFavoritesPostRequestFromJSON(json: any): ApiFavoritesPostRequest {
    return ApiFavoritesPostRequestFromJSONTyped(json, false);
}

export function ApiFavoritesPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiFavoritesPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'posterPath': json['poster_path'],
        'tmdbId': json['tmdbId'],
    };
}

export function ApiFavoritesPostRequestToJSON(json: any): ApiFavoritesPostRequest {
    return ApiFavoritesPostRequestToJSONTyped(json, false);
}

export function ApiFavoritesPostRequestToJSONTyped(value?: ApiFavoritesPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'poster_path': value['posterPath'],
        'tmdbId': value['tmdbId'],
    };
}